feat: implement JWT token validation for user and lesson API routes

+++ src/app/api/users/delete/[id]/route.ts
@@ -1,13 +1,28 @@
 import { apiResponse, apiResponseError, userService } from '@/lib'
-import { ApiResponseType } from '@/types/types'
-import { NextResponse } from 'next/server'
+import { ApiResponseType, JWTType } from '@/types/types'
+import * as jwt from 'jsonwebtoken'
+import { NextRequest, NextResponse } from 'next/server'
+const JWT_SECRET_KEY = process.env.JWT_SECRET!
 
 export async function DELETE(
-	request: Request,
+	request: NextRequest,
 	{ params }: { params: Promise<{ id: string }> }
 ): Promise<NextResponse<ApiResponseType>> {
 	const { id } = await params
+	const accessToken = request.cookies.get('accessToken')?.value
 
+	if (!accessToken) {
+		return apiResponse(
+			{ success: false, message: 'No access token provided', data: null },
+			{ status: 401 }
+		)
+	}
+
+	const isTokenValid = jwt.verify(accessToken, JWT_SECRET_KEY) as JWTType
+
+	if (!isTokenValid) {
+		return apiResponse({ success: false, message: 'Token is expired', data: null }, { status: 401 })
+	}
 	try {
 		const deletedUser = await userService.deleteById(id)
